Start modularizing wasm validation, instrumentation and execution
