+++ radix-engine/src/engine/mod.rs
@@ -4,7 +4,6 @@ mod id_validator;
 mod process;
 mod substate_receipt;
 mod track;
-mod wasm_env;
 
 pub use component_objects::*;
 pub use id_allocator::*;
@@ -12,4 +11,3 @@ pub use id_validator::*;
 pub use process::{Process, SNodeState, SystemApi};
 pub use substate_receipt::{SubstateOperation, CommitReceipt, SubstateOperationsReceipt};
 pub use track::{Track, BorrowedSNodes, SubstateUpdate, SubstateParentId};
-pub use wasm_env::{EnvModuleResolver, ENGINE_FUNCTION_INDEX, ENGINE_FUNCTION_NAME};
