+++ radix-engine/src/wasm/mod.rs
@@ -1,0 +1,12 @@
+mod env_resolver;
+mod validation;
+mod instrumentation;
+mod execution;
+mod errors;
+
+pub use env_resolver::{EnvModuleResolver, ENGINE_FUNCTION_INDEX, ENGINE_FUNCTION_NAME};
+pub use validation::{parse_module, validate_module};
+pub use execution::{instantiate_module};
+pub use errors::*;
+pub use errors::*;
+
